requirements.txt
Pillow
Requests
streamlit
streamlit_lottie
pybase64